Develop

See merge request smirnov/morris_butler!31

+++ calendars/views.py
@@ -3,16 +3,38 @@ from django.shortcuts import render
 from django.contrib.auth.decorators import *
 from datetime import datetime, timedelta
 from calendars.forms import *
+from main.models import Language
 from django.http import HttpResponse, HttpResponseRedirect
 import json
+from localisation import rus, eng
 
 
 @login_required
 def index(request):
+    """!
+       @brief This function render calendar page for current user
+    """
     if request.method == "GET":
-        date = datetime.now()
         context = dict(title="Calendar index page", header="Calendar index page header")
-        context['weeks'] = get_weeks(date)
+        user_lang = Language.objects.filter(user=request.user).first().lang
+        if user_lang == "ru":
+            lang = rus
+        elif user_lang == "en":
+            lang = eng
+        else:
+            lang = eng
+        context['language'] = user_lang
+        context['lang'] = lang
+        date = datetime.now()
+        events = Event.objects.filter(user=request.user)
+        context['weeks'] = get_weeks(date, request.user, events)
+        context['all_events_count'] = events.count()
+        context['today_events_count'] = events.filter(date=date.date()).count()
+        context['yesterday_events_count'] = events.filter(date=date.date() - timedelta(1)).count()
+        context['week_events_count'] = last_events = events.filter(date__lte=date.date())
+        context['week_events_count'] = last_events.filter(date__gte=date.date() - timedelta(1)).count()
+        context['month_events_count'] = last_events.filter(date__gte=date.date() - timedelta(30)).count()
+        context['year_events_count'] = last_events.filter(date__gte=date.date() - timedelta(365)).count()
         context['now_month'] = get_month_name(date.month)
         context['now_year'] = date.year
         context['now_month_num'] = date.month
@@ -36,14 +58,18 @@ def index(request):
                 return HttpResponse(json.dumps({'result': 'failed'}), content_type="application/json")
         except ValueError:
             return HttpResponse(json.dumps({'result': 'failed'}), content_type="application/json")
-        response_data['weeks'] = get_weeks(now_date)
+        events = Event.objects.filter(user=request.user)
+        response_data['weeks'] = get_weeks(now_date, request.user, events)
         response_data['month_name'] = get_month_name(now_date.month)
         response_data['now_year'] = now_date.year
         response_data['result'] = "100"
         return HttpResponse(json.dumps(response_data), content_type="application/json")
 
 
-def get_weeks(date_time):
+def get_weeks(date_time, user, events):
+    """!
+        @brief This function select weeks and belong them events
+    """
     date = date_time
     datetime_now = datetime.now()
     if (date.month == datetime_now.month) and (date.year == datetime_now.year):
@@ -72,8 +98,8 @@ def get_weeks(date_time):
             else:
                 if date.month != date_time.month:
                     day_class = "future-date"
+            tmp_day_events = events.filter(date=date)
             tmp_day['class'] = day_class
-            tmp_day_events = Event.objects.filter(date=date)
             if tmp_day_events.count() > 0:
                 tmp_day['event'] = {'caption': tmp_day_events[0].title, 'id': tmp_day_events[0].id}
             week_days.append(tmp_day)
@@ -98,7 +124,7 @@ def get_weeks(date_time):
                 if date.month != date_time.month:
                     day_class = "future-date"
             tmp_day['class'] = day_class
-            tmp_day_events = Event.objects.filter(date=date)
+            tmp_day_events = events.filter(date=date)
             if tmp_day_events.count() > 0:
                 tmp_day['event'] = {'caption': tmp_day_events[0].title, 'id': tmp_day_events[0].id}
             week_days.append(tmp_day)
@@ -109,12 +135,18 @@ def get_weeks(date_time):
 
 
 def get_month_name(month):
+    """!
+        @brief This functon get name of month by number
+    """
     m = ["Январь", "Февраль", "Март", "Апрель", "Май", "Июнь",
          "Июль", "Август", "Сентябрь", "Октябрь", "Ноябрь", "Декабрь", ]
     return m[month - 1]
 
 
 def search_events(string):
+    """!
+        @brief This function select events by regular phrase
+    """
     events = Event.get_events("title_up_all")
     found_events = []
     for i in events:
@@ -124,6 +156,9 @@ def search_events(string):
 
 
 def add_event(request, data):
+    """!
+        @brief This function add event to DB if data of event is valid
+    """
     time_now = datetime.now()
     data['added_date'] = time_now.date()
     data['added_time'] = time_now.time()
@@ -147,6 +182,9 @@ def add_event(request, data):
                           should_notify_hours=data['should_notify_hours'],
                           should_notify_minutes=data['should_notify_minutes'],
                           should_notify_days=data['should_notify_days'])
+            if data['place'] != "":
+                event.place = data['place']
+            print(event.place)
             event.save()
             return "100"
 
@@ -173,15 +211,19 @@ def event_view(request):
 
 @login_required
 def get_event_data_ajax(request):
+    """!
+        This function get data of event (with ajax)
+    """
     response_data = {}
     if request.method == "POST":
         event_id = request.POST.get('id')
-        if Event.objects.filter(id=event_id).count() > 0:
+        if Event.objects.filter(user=request.user, id=event_id).count() > 0:
             event = Event.objects.filter(id=event_id).first()
             response_data = {
                 'title': event.title,
                 'description': event.description,
                 'date': datetime.combine(event.date, event.time).strftime("%I:%M%p on %B %d, %Y"),
+                'map_coordinates': event.place,
             }
             result = '100'
         else:
@@ -197,8 +239,11 @@ def delete_ajax(request):
     response_data = {}
     if request.method == "POST":
         event_id = request.POST.get('id')
-        if Event.delete_event(event_id):
-            result = "100"
+        if Event.objects.filter(user=request.user, id=event_id).count() > 0:
+            if Event.delete_event(event_id):
+                result = "100"
+            else:
+                result = "114"
         else:
             result = "114"
         response_data['result'] = result
