+++ notes/views.py
@@ -1,38 +1,55 @@
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from notes.forms import *
+from main.models import Language
 import json
 from datetime import datetime
 from notes.forms import AddNoteForm, SearchForm
 from notes.models import Notes
 from django.http import HttpResponse, HttpResponseRedirect
+from django.core.paginator import Paginator
+from localisation import eng, rus
 
 
 @login_required
 def index(request):
     context = {
         'title': "Notes index page",
-        'header': "Notes index page header",
     }
-    user = request.user
+    user_lang = Language.objects.filter(user=request.user).first().lang
+    if user_lang == "ru":
+        lang = rus
+    elif user_lang == "en":
+        lang = eng
+    else:
+        lang = eng
+    context['language'] = user_lang
+    context['lang'] = lang
     notes_list = []
-    context['voice_note'] = Notes.objects.filter(user=request.user, is_voice=True).count()
-    context['text_note'] = Notes.objects.filter(user=request.user, is_voice=False).count()
-    notes = Notes.get_notes('title_up', user)
+    notes = Notes.objects.filter(user=request.user)
+    context['all_notes_count'] = notes.count()
+    context['voice_notes_count'] = notes.filter(is_voice=True).count()
+    context['text_notes_count'] = int(context['all_notes_count']) - int(context['voice_notes_count'])
+    notes = notes.order_by('name')
     for i in notes:
         notes_list.append((i.name, i.added_time.strftime("%I:%M%p on %B %d, %Y"), i.id, i.data_part))
-    context['notes_data'] = notes_list
+    pages = Paginator(notes_list, 20)
+    context['notes_data'] = pages.page(1)
+    context['notes_pages'] = pages.page_range
     context['search_note_form'] = SearchForm()
     context['add_note_form'] = AddNoteForm()
     context['edit_note_form'] = EditNoteForm()
     return render(request, "notes/index.html", context)
 
+
 @login_required
 def add_note_ajax(request):
+    """
+    @brief
+    This function adds notes
+    """
     response_data = {}
     if request.method == "POST":
-        print(request.POST)
         form = AddNoteForm(request.POST)
         if form.is_valid():
             name = form.cleaned_data['note_title']
@@ -53,9 +70,13 @@ def add_note_ajax(request):
         return HttpResponseRedirect('/')
 
 
-def search(substr):
+def search_notes(substr, user):
     obj = Notes.objects.all()
     ret_list = []
+    """
+        @brief
+        This function searches notes
+        """
     for i in obj:
         if substr in i.name:
             ret_list.append((i.name, i.added_time.strftime("%I:%M%p on %B %d, %Y"), i.id, [i.data]))
@@ -64,10 +85,16 @@ def search(substr):
 
 @login_required
 def get_note_data_ajax(request):
+    """
+    @brief
+    This function receives information about notes
+    @detailed
+    This function receives information about notes such as date, time, name
+    """
     response_data = {}
     if request.method == "POST":
         note_id = request.POST.get('id')
-        if Notes.objects.filter(id=note_id).count() > 0:
+        if Notes.objects.filter(user=request.user, id=note_id).count() > 0:
             note = Notes.objects.filter(id=note_id).first()
             response_data = {
                 'title': note.name,
@@ -87,12 +114,18 @@ def get_note_data_ajax(request):
 
 @login_required
 def search_ajax(request):
+    """
+    @brief
+    This function looks for the notes
+    @detailed
+    This function searches for a note among those that exist
+    """
     if request.method == "POST":
         form = SearchForm(request.POST)
         if form.is_valid():
             string = form.data['result']
             response_data = {
-                'notes_list': search_notes(string, request.user)
+                'notes_list': Notes.search_notes(string, request.user)
             }
             result = '100'
         else:
@@ -108,6 +141,12 @@ def search_ajax(request):
 
 @login_required
 def sort_ajax(request):
+    """
+    @brief
+    This function sorts notes
+    @detailed
+    This function sorts notes by time and alphabetically
+    """
     response_data = {}
     if request.method == "POST":
         sorting_types = ('date_up', 'date_down', 'title_up', 'title_down')
@@ -130,12 +169,18 @@ def sort_ajax(request):
 
 @login_required
 def save_ajax(request):
+    """
+    @brief
+    This function saves the notes
+    @detailed
+    This function saves the written note
+    """
     response_data = {}
     if request.method == "POST":
         form = EditNoteForm(request.POST)
         if form.is_valid():
             note_id = form.cleaned_data['note_id']
-            if Notes.objects.filter(id=note_id).count() > 0:
+            if Notes.objects.filter(user=request.user, id=note_id).count() > 0:
                 tmp = Notes.objects.filter(id=note_id).first()
                 tmp.name = form.cleaned_data['note_title_edit']
                 tmp.data = form.cleaned_data['note_data_edit']
@@ -157,12 +202,19 @@ def save_ajax(request):
 
 @login_required
 def delete_ajax(request):
+    """
+    @brief
+    This function deletes notes
+    """
     response_data = {}
     if request.method == "POST":
         note_id = request.POST.get('id')
         should_return_last_note = request.POST.get('return_last_note')
-        if Notes.delete_note(note_id):
-            result = "100"
+        if Notes.objects.filter(user=request.user, id=note_id).count() > 0:
+            if Notes.delete_note(note_id):
+                result = "100"
+            else:
+                result = "111"
         else:
             result = "111"
         last_notes = Notes.get_notes("date_up", request.user)
@@ -174,6 +226,24 @@ def delete_ajax(request):
             response_data['id'] = last_note.id
             response_data['name'] = last_note.name
         response_data['result'] = result
+        return HttpResponse(json.dumps(response_data), content_type="application/json")
+    else:
+        return HttpResponseRedirect('/')
+
+
+@login_required
+def paginate(request):
+    if request.method == "POST":
+        response_data = {}
+        notes_list = []
+        notes = Notes.objects.filter(user=request.user)
+        for i in notes:
+            notes_list.append((i.name, i.added_time.strftime("%I:%M%p on %B %d, %Y"), i.id, i.data_part))
+        pages = Paginator(notes_list, 20)
+        page_number = request.POST.get('page')
+        response_data['buttons'] = [pages.page(page_number).has_previous(), pages.page(page_number).has_next()]
+        response_data['notes_list'] = pages.page(page_number).object_list
+        response_data['result'] = 200
         return HttpResponse(json.dumps(response_data), content_type="application/json")
     else:
         return HttpResponseRedirect('/')
